% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{getAllCountries}
\alias{getAllCountries}
\title{Return indicators information from Trading Economics API}
\usage{
getAllCountries(outType = NULL)
}
\value{
Return a list or dictionary of all countries
}
\description{
Return indicators information from Trading Economics API
}
\section{Notes}{

Without credentials default information will be provided.
}

\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
