% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{getCalendarData}
\alias{getCalendarData}
\title{Return calendar events from Trading Economics API}
\usage{
getCalendarData(country = NULL, indicator = NULL, initDate = NULL,
  endDate = NULL, outType = NULL)
}
\arguments{
\item{country}{string or list.
String to get data for one country. List of strings to get data for
several countries. For example, country = c('United States', 'Australia').}

\item{indicator}{string or list.
String  to get data for one category. List of strings to get data for several calendar events.
For example, category = 'GDP Growth Rate' or
category = c('Exports', 'Imports').}

\item{initDate}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{endDate}{string with format: YYYY-MM-DD.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return list or data frame of calendar events.
}
\description{
Return calendar events from Trading Economics API
}
\section{Notes}{

All parameters are optional. When not supplying parameters, data for all countries and indicators will be provided.
Without credentials, only sample data is returned.
}
\examples{
\dontrun{ getCalendarData(country = 'United States', indicator = 'Composite Pmi',
 initDate = '2011-01-01', endDate = '2016-01-01')
getCalendarData(country = c('United States', 'India'),
 indicator = c('Composite Pmi', 'Bankruptcies'),
  initDate = '2011-01-01', endDate = '2016-01-01')
  }

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}

