% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{getCalendarUpdates}
\alias{getCalendarUpdates}
\title{----------------------------------
-------- Calendar Updates --------
----------------------------------
Return historical information from Trading Economics API.}
\usage{
getCalendarUpdates(outType = NULL)
}
\arguments{
\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return list or data frame of calendar updates.
}
\description{
----------------------------------
-------- Calendar Updates --------
----------------------------------
Return historical information from Trading Economics API.
}
\section{Notes}{

Without credentials, only sample data is returned.
}

\examples{
\dontrun{getCalendarUpdates()
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getCalendarData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
