% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtradeCountryByCategory.R
\name{getCmtCountryByCategory}
\alias{getCmtCountryByCategory}
\title{Get detailed information about Comtrade Country by Imports or Exports and by Category from Trading Economics API}
\usage{
getCmtCountryByCategory(
  country = NULL,
  type = NULL,
  category = NULL,
  outType = NULL
)
}
\arguments{
\item{country}{string.
String to get data for one country.}

\item{type}{string.
String to get data for import or export.}

\item{category}{string.
String to get data for category.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Get detailed information about Comtrade Country by Imports or Exports and by Category
}
\description{
Get detailed information about Comtrade Country by Imports or Exports and by Category from Trading Economics API
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{
getCmtCountryByCategory(country = 'Portugal', type = 'import', 
category = NULL, outType = NULL )
getCmtCountryByCategory(country = 'Brazil', type = import, 
category = 'Swine, live', outType = 'df' )
}


}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
