% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtradeCountryFilterByType.R
\name{getCmtCountryFilterByType}
\alias{getCmtCountryFilterByType}
\title{Get detailed information about Comtrade Countries filter by type 'import' or 'export' from Trading Economics API}
\usage{
getCmtCountryFilterByType(
  country1 = NULL,
  country2 = NULL,
  type = NULL,
  outType = NULL
)
}
\arguments{
\item{country1}{string.
String to get data for country1.}

\item{country2}{string.
String to get data for country2.}

\item{type}{string.
String to get data for import or export.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Get detailed information about Comtrade Countries filter by type 'import' or 'export'
}
\description{
Get detailed information about Comtrade Countries filter by type 'import' or 'export' from Trading Economics API
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{
getCmtCountryFilterByType(country1 = 'Portugal', country2 = 'Spain',
type = 'import', outType = NULL )
getCmtCountryFilterByType(country1 = 'United States', 
type = 'export', outType = 'df' )
}


}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
