% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtradeTotalType.R
\name{getCmtTotalByType}
\alias{getCmtTotalByType}
\title{Get detailed information about Comtrade Country Total by Import or Exports from Trading Economics API}
\usage{
getCmtTotalByType(country = NULL, type = NULL, outType = NULL)
}
\arguments{
\item{country}{string.
String to get data for one country.}

\item{type}{string.
String to get data for import or export.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return a list or data frame of Comtrade Total by Import or Export.
}
\description{
Get detailed information about Comtrade Country Total by Import or Exports from Trading Economics API
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{
getCmtTotalByType(country = 'Portugal', type = 'import', outType = Null )
getCmtTotalByType(country = 'United States', type = 'export', outType = NULL )
getCmtTotalByType(country = 'Brazil', type = 'import', outType = 'df' ) 
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
