% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade.R
\name{getComtradeCategories}
\alias{getComtradeCategories}
\title{Get Comtrade values from Trading Economics API}
\usage{
getComtradeCategories(outType = NULL)
}
\arguments{
\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data from World Bank.
}
\description{
Get Comtrade values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided. Without a category, a list of all categories will be provided.
}

\examples{
\dontrun{ getComtradeCategories(outType = 'df')
}
}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
