% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade.R
\name{getComtradeCountry}
\alias{getComtradeCountry}
\title{Get Comtrade values from Trading Economics API}
\usage{
getComtradeCountry(
  country = NULL,
  page_number = NULL,
  outType = NULL,
  id = NULL
)
}
\arguments{
\item{country}{string.}

\item{page_number}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data about Comtrade.
}
\description{
Get Comtrade values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided. Without a country, a list of countries will be provided.
}

\examples{
\dontrun{ getComtradeCountry(outType = 'df')
getComtradeCountry('portugal')
getComtradeCountry('portugal', page_number = '2')
getComtradeCountry(c('portugal', 'spain'), page_number = '4')
}

}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
