% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade.R
\name{getComtradeHistorical}
\alias{getComtradeHistorical}
\title{Get Comtrade values from Trading Economics API}
\usage{
getComtradeHistorical(symbol = NULL, outType = NULL)
}
\arguments{
\item{symbol}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or historical data about Comtrade.
}
\description{
Get Comtrade values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided.A symbol must be provided.
}

\examples{
\dontrun{ getComtradeHistorical('PRTESP24031')
}


}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
