% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditRating.R
\name{getCreditRating}
\alias{getCreditRating}
\title{Return ratings information from Trading Economics API}
\usage{
getCreditRating(country = NULL, outType = NULL)
}
\arguments{
\item{country}{string or list.
String for one country information. List of strings for
several countrys, for example country = c('country_name', 'country_name').}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}
}
\value{
Return a list or data frame of historical information for specific country or countries.
}
\description{
Return ratings information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list of all indicators will be provided exmp('').
Without credentials default information will be provided.
}

\examples{
\dontrun{getCreditRating('United States')
getCreditRating(country = c('United States', 'Portugal'))
getCreditRating(outType = 'df')}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
