% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditRating.R
\name{getCreditRatingHistorical}
\alias{getCreditRatingHistorical}
\title{Return credit rating historical information by country, or countries, default is United Satates}
\usage{
getCreditRatingHistorical(
  country = NULL,
  outType = NULL,
  initDate = NULL,
  endDate = NULL
)
}
\arguments{
\item{country}{string or list.
String for one country information. List of strings for
several countrys, for example country = c('country_name', 'country_name').}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}

\item{initDate}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{endDate}{string with format: YYYY-MM-DD.}
}
\value{
Return a list or data frame of historical information for specific country or countries.
}
\description{
Return credit rating historical information by country, or countries, default is United Satates
}
\section{Notes}{

country needs to be provided.
}

\examples{
\dontrun{getCreditRatingHistorical('United States')
getCreditRatingHistorical(country = c('United States', 'Portugal'))
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
