% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{getDiscontinuedIndicators}
\alias{getDiscontinuedIndicators}
\title{Returns List of Financials Discontinued Indicators by country name from Trading Economics API}
\usage{
getDiscontinuedIndicators(country = NULL, outType = NULL)
}
\arguments{
\item{country}{string or list.
String to get data for one country List of strings to get data for more than one country}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Returns List of Financials Discontinued Indicators by country name.
}
\description{
Returns List of Financials Discontinued Indicators by country name from Trading Economics API
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{
getDiscontinuedIndicators(country =c("united states","china"), outType = 'df')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
