% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnings.R
\name{getEarnings}
\alias{getEarnings}
\title{Return earnings information from Trading Economics API}
\usage{
getEarnings(
  indicator = NULL,
  initDate = NULL,
  endDate = NULL,
  type = NULL,
  category = NULL,
  outType = NULL
)
}
\arguments{
\item{indicator}{string.
string for symbol}

\item{initDate, }{string.}

\item{endDate, }{string.}

\item{type, }{string.Could be earnings, ipo, or dividends.}

\item{category}{string or list.
string for market symbol}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}
}
\value{
Returns a list or data frame of information for specific country, or for specific symbol by initial and end dates.
}
\description{
Return earnings information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list will be provided exmp('').
Without credentials default information will be provided.
}

\examples{
\dontrun{getEarnings(outType = 'df')
getEarnings('united states', outType = 'df' )
getEarnings(initDate = '2017-01-01', outType = 'df')
getEarnings(indicator='msft:us', category="symbol" , '2017-01-01' )
getEarnings(indicator='msft:us', category="symbol", '2016-01-01','2017-01-01', outType = 'df')
getEarnings(type ='dividends', outType = 'df')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
