% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\name{getEurostatCategoryGroups}
\alias{getEurostatCategoryGroups}
\title{Returns List of categories and category groups available.}
\usage{
getEurostatCategoryGroups(outType = NULL)
}
\arguments{
\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Returns List of categories and category groups available.
}
\description{
Returns List of categories and category groups available.
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{
getEurostatCategoryGroups(outType = 'df')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
