% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\name{getEurostatCountries}
\alias{getEurostatCountries}
\title{Returns List of List of countries available. from Trading Economics API}
\usage{
getEurostatCountries(outType = NULL)
}
\arguments{
\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Returns List of countries available..
}
\description{
Returns List of List of countries available. from Trading Economics API
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{
getEurostatCountries(outType = 'df')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
