% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\name{getEurostatData}
\alias{getEurostatData}
\title{Return search information from Trading Economics API}
\usage{
getEurostatData(
  lists = NULL,
  category_group = NULL,
  category = NULL,
  country = NULL,
  outType = NULL
)
}
\arguments{
\item{lists}{string.
String for eurostats lists, for example lists = 'countries' or lists='categories'.}

\item{category_group}{string.
String for one category_group information, for example category_group = 'Poverty'.}

\item{category}{string.
String for category information, for example category = 'People at risk of income poverty after social transfers'.}

\item{country}{string.
String for eurostats data by country, for example category = 'Denmark'.}

\item{outType}{string.
'lst'(default) for list format output, 'df' for data frame,}
}
\value{
Return a list or data frame of all countries or categories available and returns Eurostat data by country/category/category_group.
}
\description{
Return search information from Trading Economics API
}
\section{Notes}{

It's necessary to provide some parameter to show data.
Without credentials default information will be provided.
}

\examples{
\dontrun{ getEurostatData(lists = 'countries')
getEurostatData(lists = 'categories')
getEurostatData(country = 'Denmark')
getEurostatData(category = 'People at risk of...')
getEurostatData(category_group = 'Poverty')
getEurostatData(country = 'Denmark', category = 'People ...')
getEurostatData(country = 'Denmark', category_group = 'Poverty')
}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
