% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\name{getEurostatHistorical}
\alias{getEurostatHistorical}
\title{---------------------------------
------ EUROSTAT HISTORICAL ------
---------------------------------
Return historical information from Trading Economics API.}
\usage{
getEurostatHistorical(
  id = NULL,
  start_date = NULL,
  end_date = NULL,
  outType = NULL
)
}
\arguments{
\item{id}{string or list.
String to get data for one ID historic.}

\item{start_date}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{end_date}{string with format: YYYY-MM-DD.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return a list or data frame of historical information for specific Eurostat ID.
}
\description{
---------------------------------
------ EUROSTAT HISTORICAL ------
---------------------------------
Return historical information from Trading Economics API.
}
\section{Notes}{

An ID must be provided.
}

\examples{
\dontrun{ getEurostatHistorical(id = '24804', outType = 'df')
getEurostatHistorical(id = '24804', start_date = '2017-01-01')
getEurostatHistorical(id = '24804', start_date = '2017-01-01', end_date = '2020-05-05')

}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}} and \code{\link{getSearch}}
}
