% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalReserve.R
\name{getFedHistorical}
\alias{getFedHistorical}
\title{Get Federal Reserve values from Trading Economics API}
\usage{
getFedHistorical(symbol = NULL, outType = NULL)
}
\arguments{
\item{symbol}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or historical data from Federal Reserve.
}
\description{
Get Federal Reserve values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided. A symbol must be provided.
}

\examples{
\dontrun{ getFedHistorical(c('RACEDISPARITY005007', 'AGEXMAK2A647NCEN'))
getFedHistorical('RACEDISPARITY005007')
}

}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
