% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalReserve.R
\name{getFedReserve}
\alias{getFedReserve}
\title{Get Federal Reserve values from Trading Economics API}
\usage{
getFedReserve(indicator = NULL, category = NULL, outType = NULL)
}
\arguments{
\item{indicator}{string.}

\item{category}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data from Federal Reserve.
}
\description{
Get Federal Reserve values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided. Without a category, a list of the name of all states will be provided.
}

\examples{
\dontrun{ getFedReserve(indicator = 'arkansas', category = 'county')
getFedReserve(category = 'states')
}

}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
