% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalReserve.R
\name{getFedReserveSnapshot}
\alias{getFedReserveSnapshot}
\title{Get Federal Reserve values from Trading Economics API}
\usage{
getFedReserveSnapshot(
  symbol = NULL,
  state = NULL,
  county = NULL,
  country = NULL,
  URL = NULL,
  outType = NULL
)
}
\arguments{
\item{symbol}{string.}

\item{state}{string.}

\item{county}{string.}

\item{country}{string.}

\item{URL}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data from Federal Reserve.
}
\description{
Get Federal Reserve values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided.
}

\examples{
\dontrun{ getFedReserveSnapshot('AGEXMAK2A647NCEN', outType = 'df')
getFedReserveSnapshot(county = 'arkansas')
getFedReserveSnapshot(county ='Pike County')
getFedReserveSnapshot(country = 'united states')
getFedReserveSnapshot(country = 'united states')
getFedReserveSnapshot(URL = '/united-states/income-inequality-in-aleutians-east
-borough-ak-fed-data.html' )
}

}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
