% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts.R
\name{getForecastData}
\alias{getForecastData}
\title{Return forecast values from Trading Economics API}
\usage{
getForecastData(
  country = NULL,
  indicator = NULL,
  outType = NULL,
  ticker = NULL
)
}
\arguments{
\item{country}{string or list.
String to get data for one country. List of strings to get data for
several countries. For example, country = c('United States', 'Australia').}

\item{indicator}{string or list.
String  to get data for one category. List of strings to get data for several calendar events.
For example, category = 'GDP Growth Rate' or
category = c('Exports', 'Imports').}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list .}

\item{ticker}{string or list.}
}
\value{
Return a list or dataframe of forecast values by country, by indicator, by country and indicator.
}
\description{
Return forecast values from Trading Economics API
}
\section{Notes}{

At least one of parameters, country or indicator, should be provided.
Without credentials, only sample data is returned.
}

\examples{
\dontrun{ getForecastData(country = 'United States', indicator = 'Imports')
getForecastData(country = c('United States', 'china'), indicator = c('gdp','inflation rate'))
getForecastData(country = 'United States')
getForecastData(indicator = 'Imports')
}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getIndicatorData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
