% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicalDB.R
\name{getHistorical}
\alias{getHistorical}
\title{Return historical information from Trading Economics API.}
\usage{
getHistorical(symbol = NULL, initDate = NULL, endDate = NULL, outType = NULL)
}
\arguments{
\item{symbol}{string or list.
String to get data for one symbol. List of strings to get data for more than one symbol.}

\item{initDate}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{endDate}{string with format: YYYY-MM-DD.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return a list or data frame of historical information for specific symbol.
}
\description{
Return historical information from Trading Economics API.
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{getHistorical('aapl:us')
getHistorical('USURTOT')
getHistorical('USURTOT', '2016-01-01')
getHistorical('aapl:us', '2016-01-01'),
getHistorical('RACEDISPARITY005007:fred')
getHistorical('PRTESP24031:comtrade')
getHistorical('usa.fr.inr.rinr:worldbank')
}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
