% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{getHistoricalFinancialsData}
\alias{getHistoricalFinancialsData}
\title{Return Financials historical data by stock symbol from Trading Economics API}
\usage{
getHistoricalFinancialsData(symbol = NULL, category = NULL, outType = NULL)
}
\arguments{
\item{symbol}{string or list.
String to get data for one symbol. List of strings to get data for more than one symbol.}

\item{category}{string.
String to get data for one indicator.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return a list or data frame of Financials historical data by stock symbol.
}
\description{
Return Financials historical data by stock symbol from Trading Economics API
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{
getHistoricalFinancialData(symbol = 'aapl:us', category = 'assets', outType = 'df')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
