% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{getHistoricalMarkets}
\alias{getHistoricalMarkets}
\title{Return historical information from Trading Economics API.}
\usage{
getHistoricalMarkets(
  symbol = NULL,
  initDate = NULL,
  endDate = NULL,
  outType = NULL
)
}
\arguments{
\item{symbol}{string or list.
String to get data for one symbol. List of strings to get data for
several symbols. For example, symbol = c('AAPL:US', 'INDU:IND').}

\item{initDate}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{endDate}{string with format: YYYY-MM-DD.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return a list or data frame of historical information for specific market.
}
\description{
Return historical information from Trading Economics API.
}
\section{Notes}{

A symbol must be provided.
}

\examples{
\dontrun{ getHistoricalMarkets(c('AAPL:US','INDU:IND'),
initDate = '2011-01-01', endDate = '2016-01-01', outType = 'df')
getHistoricalMarkets('aapl:us', initDate =  '2019-02-20' )
getHistoricalMarkets('AAPL:US', initDate = '2011-01-01', endDate = '2016-01-01')

}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
