% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{getHistoricalUpdates}
\alias{getHistoricalUpdates}
\title{Return historical information from Trading Economics API.}
\usage{
getHistoricalUpdates(outType = NULL)
}
\arguments{
\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return a list or data frame of historical updates information.
}
\description{
Return historical information from Trading Economics API.
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{ getHistoricalUpdates()
}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
