% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{getIndicatorData}
\alias{getIndicatorData}
\title{Return indicators information from Trading Economics API}
\usage{
getIndicatorData(country = NULL, indicator = NULL, outType = NULL,
  credentials = NULL)
}
\arguments{
\item{country}{string or list.
String for one country information. List of strings for
several countrys, for example country = c('country_name', 'country_name').}

\item{indicator}{string or list.
String for one indicator. List of strings for several indicators, for example
indicators = 'indicator_name' or
indicators = c('indicator_name', 'indicator_name').}

\item{outType}{string.
'dict'(default) for dictionary format output, 'df' for data frame,
'raw' for list of dictionaries directly from the web.}

\item{credentials}{string.
User's credentials.}
}
\value{
Return a list or dictionary of all indicators, indicators by country or country-indicator pair.
}
\description{
Return indicators information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list of all indicators will be provided.
Without credentials default information will be provided.
}
\examples{
getIndicatorData(country = 'United States', indicators = 'Imports', output_type = 'df')
getIndicatorData(country = c('United States', 'Portugal'), indicators = c('Imports','Exports'))
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}

