% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{getLatestArticles}
\alias{getLatestArticles}
\title{Return latest articles information from Trading Economics API}
\usage{
getLatestArticles(country = NULL, indicator = NULL, id = NULL,
  limit = NULL, start = NULL, initDate = NULL, endDate = NULL,
  outType = NULL)
}
\arguments{
\item{country}{string or list.
String for one country information. List of strings for
several countrys, for example country = c('country_name', 'country_name').}

\item{indicator, }{string.
String for one indicator, or list of strings for several indicators.}

\item{id, }{string.
string for a specific Id.}

\item{limit, }{string.
Limits the list size.}

\item{start, }{string.
Specific starting index.}

\item{initDate, }{string.}

\item{endDate, }{string.}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}
}
\value{
Return a list or data frame of the latest articles, by country or countries, by indicator and with page limit and start index number..
}
\description{
Return latest articles information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list of all articles will be provided exmp('').
Without credentials default information will be provided.
}

\examples{
\dontrun{getLatestArticles( )
getLatestArticles('United States')
getLatestArticles(c('United States', 'Portugal'))
getLatestArticles(indicator = 'inflation rate')
getLatestArticles( c('United States', 'Portugal'), 'inflation rate')
getLatestArticles(limit = '15', start = '10')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
