% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{getLatestNews}
\alias{getLatestNews}
\title{Return latest news information from Trading Economics API}
\usage{
getLatestNews(
  identifier = NULL,
  category = NULL,
  limit = NULL,
  start = NULL,
  outType = NULL
)
}
\arguments{
\item{identifier}{string.}

\item{category}{string.}

\item{limit, }{string.
Limits the list size.}

\item{start, }{string.
Specific starting index.}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}
}
\value{
Return a list or data frame of the latest news, by country or countries, by indicator and with page limit and start index number.
}
\description{
Return latest news information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list of all news will be provided exmp('').
Without credentials default information will be provided.
}

\examples{
\dontrun{getLatestNews( )
getLatestNews('United States')
getLatestNews(c('United States', 'Portugal'))
getLatestNews( identifier = 'inflation rate', category = 'indicator')
getLatestNews( c('United States', 'Portugal'), 'inflation rate')
getLatestNews(limit = '15', start = '10')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
