% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{getMarketsData}
\alias{getMarketsData}
\title{Get markets values from Trading Economics API}
\usage{
getMarketsData(marketsField, outType = NULL, credentials = NULL)
}
\arguments{
\item{marketsField}{string. Takes either one of 'commodity','currency',
'index' or 'bond' as options.}

\item{outType}{string.
'df'(default) for data frame,
'raw' for list of unparsed data.}

\item{credentials}{string.
User's credentials.}
}
\value{
Returns a list or data frame of available commodities, currencies, indeces or bonds and their latest values.
}
\description{
Get markets values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided.
}
\examples{
getMarketsData(marketsField = 'index')
}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}

