% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{getMarketsForecast}
\alias{getMarketsForecast}
\title{Get latest markets forecast from Trading Economics API}
\usage{
getMarketsForecast(symbol = NULL, category = NULL, outType = NULL)
}
\arguments{
\item{symbol}{string.}

\item{category}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data frame of market forecasts by category or symbol.
}
\description{
Get latest markets forecast from Trading Economics API
}
\section{Notes}{

A country must be provided.
}

\examples{
\dontrun{ getMarketsForecast('index')
getMarketsForecast( symbol = 'BULGARIAGOVB10Y')
}
}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
