% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{getMarketsSearch}
\alias{getMarketsSearch}
\title{Get latest markets by peers and components from Trading Economics API}
\usage{
getMarketsSearch(country = NULL, category = NULL, outType = NULL)
}
\arguments{
\item{country}{string.}

\item{category}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data frame of available Takes either one of 'index' or 'markets', 'bonds', 'commodities',
by country and pagination.
}
\description{
Get latest markets by peers and components from Trading Economics API
}
\section{Notes}{

A country must be provided.
}

\examples{
\dontrun{ getMarketsSearch('united states'), getMarketsSearch( 'japan', 'index')
getMarketsSearch('united states', category = c('index' ,'markets'))
getMarketsSearch('japan', ('index', 'markets'), '2')
}
}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
