% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{getPeers}
\alias{getPeers}
\title{Return indicators peers information from Trading Economics API}
\usage{
getPeers(country = NULL, ticker = NULL, category = NULL, outType = NULL)
}
\arguments{
\item{country}{string}

\item{ticker}{string}

\item{category}{string}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}
}
\value{
Return a list or data frame of peers information for specific country, ticker or category.
}
\description{
Return indicators peers information from Trading Economics API
}
\section{Notes}{

Without credentials default information will be provided.
}

\examples{
\dontrun{getPeers(ticker ='CPI YOY', outType = 'df')
getPeers(country ='united states', outType = 'df')
getPeers(country ='united states', category ='money', outType = 'df')}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
