% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{getSearch}
\alias{getSearch}
\title{Return search information from Trading Economics API}
\usage{
getSearch(search_term = NULL, category = NULL, outType = NULL)
}
\arguments{
\item{search_term}{string.
String for one search term. For example
search_term = 'gold' or
search_term = 'japan'.}

\item{category}{string.
String for one category information, for example category = 'markets'.}

\item{outType}{string.
'lst'(default) for list format output, 'df' for data frame,}
}
\value{
Return a list or dictionary of all categories available, search by term/keyword and category, search for term/keyword in all categories available.
}
\description{
Return search information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list of all categories will be provided.
Without credentials default information will be provided.
}

\examples{
\dontrun{getSearch()
getSearch(search_term = 'gold', outType = 'df')
getSearch(search_term = 'japan', , category= 'markets', outType = 'df')
getCalendarData(ticker= c('IJCUSA','SPAINFACORD','BAHRAININFNRATE'),
initDate = '2018-01-01', endDate = '2018-03-01')
  }

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}} and \code{\link{getIndicatorData}} and \code{\link{getEurostatData}} and \code{\link{getEurostatHistorical}}
}
