% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbology.R
\name{getSymbology}
\alias{getSymbology}
\title{Return symbology information from Trading Economics API}
\usage{
getSymbology(symbol = NULL, ticker = NULL, isin = NULL, outType = NULL)
}
\arguments{
\item{symbol}{string or list.
string for market symbol}

\item{ticker}{string.
string for symbol}

\item{isin, }{string.}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}
}
\value{
Returns a list or data frame of information for specific symbol, ticker or ISIN of a stock.
}
\description{
Return symbology information from Trading Economics API
}
\section{Notes}{

Is required to have one parameter.
}

\examples{
\dontrun{getsymbology(symbol='aapl:us', outType = 'df')
getsymbology(ticker='aapl', outType = 'df')
getsymbology(isin='US0378331005', outType = 'df')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
