% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldBank.R
\name{getWorldBankHistorical}
\alias{getWorldBankHistorical}
\title{Get World Bank values from Trading Economics API}
\usage{
getWorldBankHistorical(series_code = NULL, outType = NULL)
}
\arguments{
\item{series_code}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data from World Bank.
}
\description{
Get World Bank values from Trading Economics API
}
\section{Notes}{

Series_code must be provided.
}

\examples{
\dontrun{ getWorldBankHistorical('usa.fr.inr.rinr')

}

}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
