% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldBank.R
\name{getWorldBankIndicators}
\alias{getWorldBankIndicators}
\title{Get World Bank values from Trading Economics API}
\usage{
getWorldBankIndicators(
  category = NULL,
  identifier = NULL,
  series_code = NULL,
  URL = NULL,
  outType = NULL
)
}
\arguments{
\item{category}{string.}

\item{identifier}{string.}

\item{series_code}{string.}

\item{URL}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data from World Bank.
}
\description{
Get World Bank values from Trading Economics API
}
\section{Notes}{

Detailed information about specific indicator for all countries using a series code , for a specific country with pagination, for a specific
indicator and country by using series_code or url.
}

\examples{
\dontrun{ getWorldBankIndicators(series_code = 'fr.inr.rinr', outType = 'df')
getWorldBankIndicators(identifier = 'portugal',category = 'country', outType = 'df')
getWorldBankIndicators(series_code = 'usa.fr.inr.rinr')
getWorldBankIndicators(URL = '/united-states/real-interest-rate-percent-wb-data.html')
}
}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
