% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalReserve.R
\name{getFedHistorical}
\alias{getFedHistorical}
\title{Get Federal Reserve values from Trading Economics API}
\usage{
getFedHistorical(
  symbol = NULL,
  initDate = NULL,
  endDate = NULL,
  outType = NULL
)
}
\arguments{
\item{symbol}{string.}

\item{initDate}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{endDate}{string with format: YYYY-MM-DD.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or historical data from Federal Reserve.
}
\description{
Get Federal Reserve values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided. A symbol must be provided.
}

\examples{
\dontrun{ getFedHistorical('RACEDISPARITY005007', outType = 'df')
getFedHistorical(c('RACEDISPARITY005007','2020RATIO002013'), outType = 'df')
getFedHistorical(symbol='BAMLC0A1CAAAEY', initDate = '2018-05-01'  , outType = 'df')
getFedHistorical(symbol='BAMLC0A1CAAAEY', initDate = '2018-05-01', endDate = '2019-01-01'  , outType = 'df'))
}

}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
