% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{getHistoricalData}
\alias{getHistoricalData}
\title{Return historical information from Trading Economics API.}
\usage{
getHistoricalData(
  country = NULL,
  indicator = NULL,
  ticker = NULL,
  initDate = NULL,
  endDate = NULL,
  outType = NULL
)
}
\arguments{
\item{country}{string or list.
String to get data for one country. List of strings to get data for
several countries. For example, country = c('United States', 'Australia').}

\item{indicator}{string or list.
String  to get data for one category. List of strings to get data for several calendar events.
For example, category = 'GDP Growth Rate' or
category = c('Exports', 'Imports').}

\item{ticker}{string.
String Unique ticker used by Trading Economics}

\item{initDate}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{endDate}{string with format: YYYY-MM-DD.}

\item{outType}{string.
'df' for data frame,
'lst'(default) for list.}
}
\value{
Return a list or data frame of historical information for specific country and indicator.
}
\description{
Return historical information from Trading Economics API.
}
\section{Notes}{

Without credentials only sample data will be provided.
}

\examples{
\dontrun{ getHistoricalData(country = 'United Kingdom', indicator = 'Imports',
initDate = '2011-01-01', endDate = '2016-01-01')
getHistoricalData(country = c('United States', 'United Kingdom'),
indicator = c('Imports','Exports'),initDate = '2011-01-01', endDate = '2016-01-01')
getHistoricalData(ticker = 'USURTOT', initDate = '2015-03-01')
getHistoricalData(country = c('United States', 'United Kingdom'),
indicator = c('Imports','Exports'))
getHistoricalData(country = 'United States', indicator = 'Imports')

}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getCalendarData}} and \code{\link{getIndicatorData}}
}
