% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{getIndicatorData}
\alias{getIndicatorData}
\title{Return indicators information from Trading Economics API}
\usage{
getIndicatorData(
  country = NULL,
  indicator = NULL,
  outType = NULL,
  ticker = NULL,
  group = NULL
)
}
\arguments{
\item{country}{string or list.
String for one country information. List of strings for
several countrys, for example country = c('country_name', 'country_name').}

\item{indicator}{string or list.
String for one indicator. List of strings for several indicators, for example
indicators = 'indicator_name' or
indicators = c('indicator_name', 'indicator_name').}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}

\item{ticker}{string or list.
String for one ticker information. List of strings for
several tickers, for example ticker = c('ticker_name', 'ticker_name').}

\item{group}{string or list.
String for one group.}
}
\value{
Return a list or dictionary of all indicators, indicators by country or country-indicator pair.
}
\description{
Return indicators information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list of all indicators will be provided.
Without credentials default information will be provided.
}

\examples{
\dontrun{ getIndicatorData(country = 'United States', indicator = 'Imports', outType = 'df')
getIndicatorData(country = c('United States', 'china'), indicator = c('gdp','inflation rate'))
getIndicatorData(country = 'United States', outType = 'df')
}
}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
