% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{getLatestNews}
\alias{getLatestNews}
\title{Return latest news information from Trading Economics API}
\usage{
getLatestNews(
  country = NULL,
  indicator = NULL,
  limit = NULL,
  start = NULL,
  initDate = NULL,
  endDate = NULL,
  outType = NULL
)
}
\arguments{
\item{country}{string.}

\item{indicator}{stirng.}

\item{limit, }{string.
Limits the list size.}

\item{start, }{string.
Specific starting index.}

\item{initDate}{string with format: YYYY-MM-DD.
For example: '2011-01-01'.}

\item{endDate}{string with format: YYYY-MM-DD.}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}
}
\value{
Return a list or data frame of the latest news, by country or countries, by indicator and with page limit and start index number.
}
\description{
Return latest news information from Trading Economics API
}
\section{Notes}{

All parameters are optional. Without parameters a list of all news will be provided exmp('').
Without credentials default information will be provided.
}

\examples{
\dontrun{getLatestNews(outType = 'df')
getLatestNews(initDate = '2021-02-02', endDate = '2021-03-03', outType = 'df')
getLatestNews(country = c('united states','china') ,outType = 'df')
getLatestNews(country = 'united states', initDate = '2021-02-02', endDate = '2021-03-03' ,outType = 'df')
getLatestNews(indicator = 'inflation rate' ,outType = 'df')
getLatestNews(indicator = 'inflation rate' , initDate = '2021-02-02', endDate = '2021-03-03' ,outType = 'df')
getLatestNews(country = 'united states' ,indicator = 'inflation rate' ,outType = 'df')
getLatestNews(country = 'united states' ,indicator = 'inflation rate', initDate = '2021-02-02', endDate = '2021-03-03', outType = 'df')
getLatestNews(limit='2',start = '1', outType = 'df')
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
