% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{getLatestUpdates}
\alias{getLatestUpdates}
\title{Return the latest updates}
\usage{
getLatestUpdates(initDate = NULL, outType = NULL, country = NULL, time = NULL)
}
\arguments{
\item{initDate}{string
String for historical data lastupdate, or for historical data per date.}

\item{outType}{string.
'lst'(default) for lis format output, 'df' for data frame,}

\item{country}{string.}

\item{time}{string.}
}
\value{
Return a list or data frame of historical information by lastupdate or by specific date.
}
\description{
Return the latest updates
}
\section{Notes}{

Historical data Symbol lastupdates.
}

\examples{
\dontrun{getLatestUpdates()
getLatestUpdates(initDate = '2018-02-22')
getLatestUpdates(country = 'united states')
getLatestUpdates(initDate = '2021-10-22', time = '15:20')  
}

}
\seealso{
\code{\link{getMarketsData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getCalendarData}}
}
