% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{getMarketsData}
\alias{getMarketsData}
\title{Get markets values from Trading Economics API}
\usage{
getMarketsData(
  marketsField = NULL,
  country = NULL,
  symbol = NULL,
  cross = NULL,
  outType = NULL
)
}
\arguments{
\item{marketsField}{string. Takes either one of 'commodity','currency',
'index' or 'bond' as options.}

\item{country}{string.}

\item{symbol}{string.}

\item{cross}{string.}

\item{outType}{string.
'df' for data frame,
'raw'(default) for list of unparsed data.}
}
\value{
Returns a list or data frame of available commodities, currencies, indeces or bonds and their latest values.
}
\description{
Get markets values from Trading Economics API
}
\section{Notes}{

Without credentials only sample information will be provided.
}

\examples{
\dontrun{ getMarketsData(marketsField = 'currency')
}
}
\seealso{
\code{\link{getCalendarData}}, \code{\link{getForecastData}}, \code{\link{getHistoricalData}} and \code{\link{getIndicatorData}}
}
